Reached a checkpoint
