+++ src/components/Footer/Footer.js
@@ -1,6 +1,6 @@
-import React, { useState, useEffect, Component } from 'react';
 import { Box, Typography, Container, Grid } from "@mui/material";
 import FooterCard from './FooterCard';
+import FooterQuotes from "./Footer-quotes";
 import wellnessImg from './../../assets/wellness2.jpg'
 
 function Footer() {
@@ -38,31 +38,7 @@ function Footer() {
 
     ]
 
-    const [quote, setQuote] = useState('');
-
-    // Sample array of quotes related to mental health therapy
-    const quotes = [
-        "You are not alone in this journey.",
-        "It's okay to ask for help when you need it.",
-        "Healing is not linear; it's okay to have setbacks.",
-        "Your mental health matters just as much as your physical health.",
-        "Self-care is not selfish; it's necessary for your well-being.",
-        "You have the strength to overcome any challenge you face.",
-        "Your feelings are valid, and it's okay to express them.",
-        "Every step forward, no matter how small, is a victory."
-    ];
-
-    // Function to get a random quote
-    const getRandomQuote = () => {
-        const randomIndex = Math.floor(Math.random() * quotes.length);
-        return quotes[randomIndex];
-    }
-
-    useEffect(() => {
-        // Fetch a random quote when the component mounts
-        setQuote(getRandomQuote());
-    }, []);
-
+    
     /* seperate the quote into a different Component */
 
     return (
@@ -112,7 +88,7 @@ function Footer() {
                 </Grid>
 
                     <Box sx={{ marginTop: '40px', textAlign: 'center' }}>
-                        <Typography variant="body1">"{quote}"</Typography>
+                        <FooterQuotes/>
                         <Typography variant="caption">© 2024 Jenkins Creative Counseling Center. All rights reserved.</Typography>
                     </Box>
             </Container>
